Split network initializer (+ refactoring)
